### 2019-10-22 #1826 @mkearney
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/302, @ropensci/editors
