### 2019-10-22 #1826 @kthyng
> Hi @arokem! Would you be willing to edit this submission? Thanks in advance.

Just kidding, I will edit it since it is from rOpenSci.
