### 2019-10-22 #1826 @kthyng
Hi @mkearney! Thanks for your submission. I am not familiar with rOpenSci, but in looking at your review there, I see that it is still open. Does that mean that it has not been completed yet?
