### 2019-10-22 #1826 @mkearney
@kthyng The review of my submission has been completed, but they may consider the joss paper as part of the process? I know @ropensci has a history and documented workflow for submitting the accompanied paper.md, but I cannot speak to the nature or extent of the relationship between the two–for that hopefully @sckott can chime in!
