### 2019-10-22 #1826 @mkearney
Nevermind, the issue has actually been closed!
