### 2019-10-22 #1826 @sckott
the ropensci review is done ✅
