### 2019-10-22 #1829 @whedon
[REVIEW]: rtweet: Collecting and analyzing Twitter data
**Submitting author:** @mkearney (<a href="http://orcid.org/0000-0002-0730-4694">Michael Wayne Kearney</a>)
**Repository:** <a href="https://github.com/ropensci/rtweet" target ="_blank">https://github.com/ropensci/rtweet</a>
**Version:** v0.7.0
