### 2019-10-22 #1829 @kthyng
This submission was already [reviewed](https://github.com/ropensci/software-review/issues/302) as part of rOpenSci.
