### 2019-10-22 #1829 @kthyng
@mkearney There is a reference that's not coming through in your paper — can you check on that?

Also I will be needing your Zenodo archive.
