### 2019-10-22 3e4c78c
Fix bibref in paper.md
