+++ paper.md
@@ -62,7 +62,7 @@ featured in numerous popular press [e.g.,
 [e.g.,
 @bossetta2018simulated; @bradley2019major; @buscema2018media;
 @erlandsen2018twitter; @gitto2019brand; @kearney2019analyzing;
-@kearney2018analyzing; @li2018sentiment; @lutkenhaus2019tailoring;
+@kearney2018analyzing; @li2019sentiment; @lutkenhaus2019tailoring;
 @lutkenhaus2019mapping; @molyneux2018media; @tsoi2018can; @unsihuay2018topic;
 @valls2017urban; @wu2018finding].
 
