### 2019-10-22 #1829 @mkearney
@kthyng I believe I fixed the missing reference issue. And is this what you need for the zenodo doi/archive: http://doi.org/10.5281/zenodo.3515425?
