### 2019-10-23 #1829 @kthyng
@mkearney Yes the reference issue looks resolved now.

Yes, that zenodo archive is what I need, though can you change the title so that it matches your paper?
