### 2019-10-24 #1829 @kthyng
Ok I think I have everything now! I can go through the acceptance process now.
